improved responsive styling
